php quadro pop. flutuante corrigido

+++ migra.sql
@@ -17,7 +17,7 @@ CREATE TABLE sys_medicao(
 	user VARCHAR(50) NOT NULL,
 	data_leitura DATE,
 	criado_em DATETIME NOW(),
-	leitura INT(11) NOT NULL,
+	leitura DECIMAL(10,2) NOT NULL,
 	pop_flut INT(11) NOT NULL DEFAULT 0,
 	permanencia INT(11) NOT NULL DEFAULT 0,
 	obs VARCHAR(255),
